IMPALA-11945: Fix Flaky Test in JwtHttpTest
The JwtHttpTest.testJwtAuthWithUntrustedJwksHttpsUrl test has been
flaky in the impala-asf-master-core Jenkins build. This build was
executed twice with this test passing in one run and failing in the
other.

This test and another test named
testJwtAuthWithTrustedJwksHttpsUrlInvalidCN both follow the same
pattern. They attempt to start up a cluster with incorrectly
configured Impala daemons. Then, the Impala coordinator's log file
is searched to ensure the coordinator failed to start up for the
expected reason.  Both these tests were introduced by IMPALA-11922.

The theory for the test flakiness is that the Impala coordinator logs
did not finish writing to disk before they were checked for the
expected startup failure message. The evidence backing this theory
is that all expected log messages were present in the log file except
for the final log line containing the error message.

Three changes were made to fix the flakiness
1. The tests where log files are search for startup failure messages
   now use a single coordinator. Only a single coordinator is needed
   thus this change eliminates potential issues caused by multiple
   coordinators.
2. The tests sleep 5 seconds before searching the log files to give
   time for the logs to be fully written to disk.
3. Log buffering in the Impala daemons was turned off by setting the
   logbuflevel startup flag to -1 which turns off in-memory log
   buffering and writes the logs directly to disk.

Change-Id: I26d8c2cffa7bc095b81eb572709f55b04ac1fa67