IMPALA-12163: Fixes two issues when outputting RPC details.
The end time of the exact same rpc call was different between stdout
and the rpc details file because the end time was calculated each
time the details were written out instead of calculating the end time
once and reusing that value.

The duration of each rpc call was being calculated incorrectly.

Change-Id: Ifd9dec189d0f6fb8713fb1c7b2b6c663e492ef05