IMPALA-12426: Switches the duration fields to be stored in decimal seconds.
The original implementation of the completed queries table
stored durations in integer nanoseconds. This change
modifies the duration fields to be stored as seconds with
up to three digits of millisecond precision.

Also reduces the default max number of queued queries to a
number that will not consume as much memory.

Existing sys.impala_query_log tables will need to be
dropped.

Testing was accomplished by modifying the python custom
cluster tests.

Change-Id: I842951a132b7b8eadccb09a3674f4c34ac42ff1b