IMPALA-13505: Fix NPE in Calcite Planner
Fixes the NullPointerException occurring when using the Calcite
planner with
test_tpcds_queries.py::TestTpcdsDecimalV2Query::test_tpcds_q8.
The NPE was thrown from the Planner where it generates the list of
columns in the query for use in the profile and workload management.

Testing was accomplished by manually running the impacted the test
and with a new custom cluster test that replicates the failing test.

Change-Id: I4d282120e596fd39a569d1ce9b25024f4f174dd0