IMPALA-13536: Workload Management Tests Failing on Init Check
Most of the workload management tests verify that the workload
management process has successfully completed. Part of this
verification ensures a catalog update has propagated the workload
management changes to the coordinators by determining the catalog
version, from the catalogd logs, that contains the workload
management table changes and ensuring that version is in the
coordinator logs.

The test flakiness occurs when multiple catalogd versions are
combined into a later version. Specifically, tests were failing
because the coordinator logs were checked for catalog version X but
the actual version in the coordinator logs was X+1.

The fix for the test flakiness is to allow for the exepected catalog
version or any later version.

Change-Id: I9f20a149ab1f45ee3506f098f8594965a24a89d3