IMPALA-13201: Remove Unused Parameter from Test retry Function
The custom cluster tests utilize the retry() function defined in
retry.py. This function takes as input another function to do the
assertions. This assertion function used to have a single boolean
parameter indicating if the retry was on its last attempt. In
actuality, this boolean was not used and thus caused flake8 failures.

This change removes this unused parameter from the assertion function
passed in to the retry function.

Change-Id: I1bce9417b603faea7233c70bde3816beed45539e