IMPALA-13885: Add Query Id to Workload Management Insert Logs
When the workload management code inserts rows into the completed
queries table, it logs an entry with information on that insert.
When impalad is compiled in debug mode, the query id is included in
the log line, but when impalad is compiled in release mode the query
id is not included in the log line. This lack of query id causes
extra steps to trace cases where the insert DML failed.

This change adds the query id to both the success and failure log
messages from the workload management main processing loop.

Additional minor cleanup was done to provide error messages on python
custom cluster test asserts, add additional asserts for the updated
log messages, and to restore a trailing newline in the workload
management flags definition file.

All test_query_log.py tests passed locally. These tests were the only
tests that asserted the log messages that were modified.

Change-Id: I3c0816f9eb6bac8c891fd0e249de8863115bf466