IMPALA-14025: Fix Failing Iceberg Rest Catalog Java Test
The test "test_rest_catalog_basic" in test_iceberg_rest_catalog.py
is failing when the target filesystem is S3.  The specific failure is
class not found org.apache.hadoop.fs.s3a.S3AFileSystem.

This patch adds the hadoop-aws to the iceberg rest catalog tests when
S3 is the target filesystem.

Change-Id: I4025242d5c00efb6acfe5f55d9885a968bb8aa8c