IMPALA-14403: Fix OpenTelemetry TLS Detection
Fixes the code that detects if the OpenTelemetry collector is using
TLS. Previously, the code only worked properly when the collector URL
was all lowercase. Also removes unnecessary checks that could cause
TLS to be enabled even when the collector URL scheme was not https.

Testing accomplished by adding new ctest tests.

Change-Id: I3bf74f1353545d280575cdb94cf135e55c580ec7