IMPALA-13679: Fix ASAN Failure in Workload Management Tests
The TickerSecondsBool class defined in ticker.h is used to
process queued completed queries at specified intervals.
This class references variables defined outside of it. The
workload management code never properly shuts down this
ticker resulting in it sometimes, during shutdown, to
access variables that have already been released.

Adds a function to the Ticker class to cause it to shut
down immediately.

Updates the workload management shutdown process to call
this function to shut down and join its Ticker. This
function call ensures the Ticker has completely shut down
before the variables that were passed to its constructor
are released.

In the workload-management-worker.cc file, adds the static
keyword to variables that are intended to be used only in
that file.

Testing
* ctest ticker-test.cc passed locally
* workload management custom cluster tests passed locally

Change-Id: I9a29f4f788cf7259c263a17272e86acdda5caa77